refactor(cloudbackupschedule): Cover validation errors with PlanOnly acc tests
